fix: #10 #130 유효성 검사를 위해 ssl 인증 추가
